[Refactor] #100 관광정보 API PlaceId Type Int -> Long 변경

+++ DaOnGil/presentation/src/main/java/kr/tekit/lion/presentation/main/adapter/viewholder/PlaceHighViewHolder.kt
@@ -18,7 +18,7 @@ class PlaceHighViewHolder(
     fun bind(item: PlaceModel) {
         with(binding) {
             root.setOnClickListener {
-                onSelectPlace(item.placeId.toLong())
+                onSelectPlace(item.placeId)
             }
 
             tvName.text = item.placeName
